agentbus/pyproject.toml
[project]
name = "agentbus-core"
version = "1.0.0"
description = "AgentBus AI Programming Assistant - Python Implementation"
authors = [
    {name = "AgentBus Team"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.12.0",
    "httpx>=0.26.0",
    "asyncssh>=2.14.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "python-json-logger>=2.0.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
    "sse-starlette>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["kode_core*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
